% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_label_from_highratio_to_lowratio.R
\name{change_label_from_highratio_to_lowratio}
\alias{change_label_from_highratio_to_lowratio}
\title{Reassign Cluster Labels Based on Outcome Ratios}
\usage{
change_label_from_highratio_to_lowratio(args, oldlabel, data_train)
}
\arguments{
\item{args}{A list of arguments, which should include the following component:
\itemize{
\item \code{K_clusters}: An integer specifying the number of clusters.
}}

\item{oldlabel}{An integer vector of the original cluster labels assigned to each sample.}

\item{data_train}{A list representing the training data. This list must include the following components:
\itemize{
\item \code{data_v}: A matrix or data frame of additional covariates.
\item \code{data_y}: A binary response vector with values 0 or 1, representing the outcome variable.
}}
}
\value{
A list containing:
\itemize{
\item \code{new_list_c}: A tensor of the new cluster labels, with clusters relabeled according to the outcome ratios.
\item \code{order_c_map}: A named vector that maps the old cluster labels to the new ones, where the names represent the original labels and the values represent the new labels.
}
}
\description{
This function reassigns cluster labels such that clusters with higher ratios of positive outcomes are assigned lower labels, and those with lower ratios are assigned higher labels. This relabeling can be useful for aligning cluster labels with the significance of outcomes.
}
\details{
The \code{change_label_from_highratio_to_lowratio} function performs the following steps:
\enumerate{
\item Calculates the count of samples and the count of positive outcomes (where \code{data_y} equals 1) for each cluster.
\item Computes the ratio of positive outcomes within each cluster.
\item Orders the clusters by their outcome ratios in descending order.
\item Reassigns cluster labels such that clusters with higher outcome ratios receive lower labels.
\item Returns the new cluster labels and the mapping from old labels to new labels.
}
}
