% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_args.R
\name{parse_args}
\alias{parse_args}
\title{Parse Command-Line Arguments for PPD-Aware Clustering}
\usage{
parse_args()
}
\value{
A list of parsed command-line arguments, each corresponding to a specific parameter required by the clustering model. This list is typically used to configure the model training and evaluation process.
}
\description{
This function parses command-line arguments for a script that performs PPD-aware clustering, setting up various parameters necessary for training and evaluating a model.
}
\details{
The \code{parse_args} function uses the \code{argparser} package to define and parse a series of command-line arguments. These arguments are essential for configuring various aspects of the PPD-aware clustering process, including model architecture, training settings, and input/output paths.

The following command-line arguments are supported:
\itemize{
\item \code{--init_AE_epoch}: Integer. Number of epochs for representation initialization.
\item \code{--n_hidden_fea}: Integer. Number of hidden units in the LSTM layers.
\item \code{--output_path}: Character. Location where output results will be saved.
\item \code{--input_path}: Character. Location of the input dataset.
\item \code{--filename_train}: Character. Path to the training data file.
\item \code{--filename_test}: Character. Path to the test data file.
\item \code{--n_input_fea}: Integer. Number of original input features.
\item \code{--n_dummy_demov_fea}: Integer. Number of dummy demographic features.
\item \code{--lstm_layer}: Integer. Number of layers in the LSTM. Default is 1.
\item \code{--lr}: Numeric. Learning rate for training. Default is 1e-4.
\item \code{--lstm_dropout}: Numeric. Dropout rate for the LSTM layers. Default is 0.0.
\item \code{--K_clusters}: Integer. Number of initial clusters.
\item \code{--iter}: Integer. Maximum number of iterations for merging clusters. Default is 20.
\item \code{--epoch_in_iter}: Integer. Number of epochs per iteration when merging clusters. Default is 1.
\item \code{--seed}: Integer. Random seed for reproducibility. Default is 1111.
\item \code{--cuda}: Integer. Flag indicating whether to use CUDA (GPU acceleration). Default is 1.
\item \code{--lambda_AE}: Numeric. Regularization weight for the autoencoder loss in each iteration. Default is 1.0.
\item \code{--lambda_classifier}: Numeric. Regularization weight for the classifier loss in each iteration. Default is 1.0.
\item \code{--lambda_outcome}: Numeric. Regularization weight for the outcome loss in each iteration. Default is 1.0.
\item \code{--lambda_p_value}: Numeric. Regularization weight for the p-value adjustment in each iteration. Default is 1.0.
}
}
